pyproject.toml
[tool.poetry]
name = "agmrv_gendiff"
version = "0.5.3"
description = "The Difference Generator"
authors = ["agmrv <airatgum@gmail.com>"]
packages = [
    { include = "gendiff" },
]
classifiers = [
  "Environment :: Console",
  "Intended Audience :: End Users/Desktop",
  "Natural Language :: English",
  "Topic :: Utilities",
]
repository = "https://github.com/agmrv/python-project-lvl2"

[tool.poetry.dependencies]
python = "^3.6"
PyYAML = "^5.3.1"

[tool.poetry.dev-dependencies]
wemake-python-styleguide = "^0.14.1"
pytest = "^6.0.2"
pytest-cov = "^2.10.1"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
